# perf(rust): prealloc decompression buffers