To avoid reallocs as decompressors write into them. Simple 4-line change that boosts decompressed throughput by +40-50%

### Bench delta

```
mcap_read/MessageStream_1M_uncompressed
                        time:   [66.406 ms 66.638 ms 66.810 ms]
                        thrpt:  [14.968 Melem/s 15.006 Melem/s 15.059 Melem/s]
                 change:
                        time:   [-4.2425% -3.4304% -2.6572%] (p = 0.00 < 0.05)
                        thrpt:  [+2.7297% +3.5523% +4.4305%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
mcap_read/MessageStream_1M_lz4
                        time:   [223.21 ms 224.45 ms 225.67 ms]
                        thrpt:  [4.4313 Melem/s 4.4554 Melem/s 4.4801 Melem/s]
                 change:
                        time:   [-36.047% -34.784% -33.497%] (p = 0.00 < 0.05)
                        thrpt:  [+50.370% +53.337% +56.364%]
                        Performance has improved.
mcap_read/MessageStream_1M_zstd
                        time:   [274.52 ms 276.51 ms 278.58 ms]
                        thrpt:  [3.5896 Melem/s 3.6165 Melem/s 3.6427 Melem/s]
                 change:
                        time:   [-28.976% -28.093% -27.277%] (p = 0.00 < 0.05)
                        thrpt:  [+37.508% +39.069% +40.797%]
                        Performance has improved.
```